+++ AngularClient/src/main/account/login.component.ts
@@ -6,8 +6,7 @@ import { timer as observableTimer } from "rxjs";
 
 @Component({
   selector: "login",
-  templateUrl: "login.component.html",
-  styleUrls: ["login.component.css"]
+  templateUrl: "login.component.html"
 })
 export class LoginComponent implements OnInit {