+++ AngularClient/src/main/admin/admin.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
 import { AppEntityManager, AppHttpClient, AuthService, User, Project } from "@forcrowd/backbone-client-core";
-import { EntityQuery } from "breeze-client";
+import { EntityQuery, Predicate } from "breeze-client";
 import { Observable } from "rxjs";
 import { finalize, mergeMap, map } from "rxjs/operators";
 
@@ -41,16 +41,27 @@ export class AdminService {
     return this.appEntityManager.executeQueryObservable<Project>(query, forceRefresh);
   }
 
-  getUserCount() {
-
-    const query = EntityQuery
+  getUser() {
+    let query = EntityQuery
       .from("Users")
-      .take(0)
-      .inlineCount(true);
+      .inlineCount(true)
+      .orderByDesc("CreatedOn");
 
     return this.appEntityManager.executeQueryObservable<User>(query).pipe(
       map((response) => {
-        return response.count;
+        return response;
+      }));
+  }
+
+  getProject() {
+    let query = EntityQuery
+      .from("Project")
+      .inlineCount(true)
+      .orderByDesc("CreatedOn");
+
+    return this.appEntityManager.executeQueryObservable<Project>(query).pipe(
+      map((response) => {
+        return response;
       }));
   }