+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Changelog
 
+**0.9.24**
+
+* AngularClient - Admin Overview updates
+* AngularClient - Minor fix for remove functions in project manager
+
 **0.9.23**
 
 * Angular - Library updates